pyproject.toml
[tool.poetry]
name = "langchain-exercise"
version = "0.1.0"
description = ""
authors = ["Jayden <jayden@deepsearch.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.13"
jupyter = "^1.0.0"
openai = "^1.14.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.1"
langchain-openai = "^0.1.1"
streamlit = "^1.32.2"
huggingface-hub = "^0.22.1"
pypdf = "^4.1.0"
faiss-cpu = "^1.8.0"
sentence-transformers = "^2.6.1"
tiktoken = "^0.6.0"
wikipedia = "^1.4.0"
numexpr = "^2.9.0"
langchain-community = "^0.0.31"
chromadb = "^0.4.24"
pypdf2 = "^3.0.1"
streamlit-chat = "^0.1.1"
finance-datareader = "^0.9.90"
plotly = "^5.20.0"
pykrx = "^1.0.45"
langchain-experimental = "^0.0.56"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
